rename rocks_db to rocksdb and segregate into feature
