+++ bothan-core/Cargo.toml
@@ -39,4 +39,9 @@ ed25519 = "2.2.3"
 ed25519-dalek = { version = "2.1.1", features = ["std", "rand_core"] }
 hex = "0.4.3"
 mini-moka = "0.10.3"
-rust-rocksdb = "0.36.0"
+rust-rocksdb = { version = "0.36.0", optional = true }
+
+[features]
+default = []
+rocksdb = ["dep:rust-rocksdb"]
+all = ["rocksdb"]
