+++ bothan-core/src/store/rocksdb/error.rs
